feat: 옮겨진 파일 추가수정 (SCRUM-166)
